+++ core/include/cubos/core/reflection/type.hpp
@@ -30,6 +30,14 @@ namespace cubos::core::reflection
         /// @return Reference to the type.
         static Type& create(std::string name);
 
+        /// @brief Constructs an unnamed type with the given identifier.
+        ///
+        /// The type is also marked as not being implemented.
+        ///
+        /// @param id Type identifier.
+        /// @return Reference to the type.
+        static Type& unnamed(unsigned long id);
+
         /// @brief Destroys the given type.
         /// @param type Type to destroy.
         static void destroy(Type& type);
@@ -87,6 +95,9 @@ namespace cubos::core::reflection
         /// @return Whether the objects have the same address, which indicates equality.
         bool operator==(const Type& other) const;
 
+        /// @brief Checks whether the type had reflection implemented for it.
+        bool implemented() const;
+
     private:
         ~Type();
 
@@ -137,6 +148,7 @@ namespace cubos::core::reflection
         };
 
         std::string mName;
+        bool mImplemented{true};
         std::vector<Trait> mTraits;
     };
 } // namespace cubos::core::reflection