+++ core/tests/reflection/reflect.cpp
@@ -38,16 +38,27 @@ CUBOS_REFLECT_EXTERNAL_TEMPLATE((typename T), (Templated<T>))
     return Type::create("Templated<" + reflect<T>().name() + ">");
 }
 
+/// @brief Type without reflection.
+struct Unreflected
+{
+};
+
 TEST_CASE("reflection::reflect")
 {
+    CHECK(reflect<Internal>().implemented());
     CHECK(reflect<Internal>().name() == "Internal");
     CHECK(reflect<Internal>().is<Internal>());
 
+    CHECK(reflect<External>().implemented());
     CHECK(reflect<External>().name() == "External");
     CHECK(reflect<External>().is<External>());
     CHECK_FALSE(reflect<External>().is<Internal>());
 
+    CHECK(reflect<Templated<Internal>>().implemented());
     CHECK(reflect<Templated<Internal>>().name() == "Templated<Internal>");
     CHECK(reflect<Templated<External>>().name() == "Templated<External>");
     CHECK(reflect<Templated<Templated<Internal>>>().name() == "Templated<Templated<Internal>>");
+
+    CHECK_FALSE(reflect<Unreflected>().implemented());
+    CHECK(reflect<Unreflected>().is<Unreflected>());
 }