+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rotated box Gizmo (#878, **@DiogoMendonc-a**).
 - Shader asset and bridge (#1058, **@tomas7770**).
 
+### Changed
+
+- Make reflection work for all types, even those without reflection implemented (#1092, **@RiscadoA**).
+
 ### Fixed
 
 - Crash in multiple samples due to missing plugin dependencies (**@RiscadoA**).